add rating and edit comments & Organize project files and views
