--- management/templates/detail_doctor.html
@@ -1,91 +1,0 @@
-{% extends 'base.html' %}
-{% load static %}
-{% block content %}
-<body>
-    {% block extra_css %}
-    <link rel="stylesheet" href="{% static 'css/detail_doctor.css' %}">
-    {% endblock %}
-    <div class="message-container">
-        {% if messages %}
-        <ul class="messages">
-            {% for message in messages %}
-                <li class="{{ message.tags }}" id="message-{{ forloop.counter }}">{{ message }}</li>
-            {% endfor %}
-        </ul>
-        {% endif %}
-    </div>
-    <div class="doctor-profile">
-        <div class="header">
-            <h1 class="doctor-name">{{ doctor.first_name }} {{ doctor.last_name }}</h1>
-            <p class="specialization"><strong>تخصص:</strong> {{ doctor.specializations }}</p>
-        </div>
-
-        <div class="contact-info">
-            <p><strong>شماره تماس:</strong> {{ doctor.phone }}</p>
-            <p><strong>آدرس مطب:</strong> {{ doctor.clinic_address }}</p>
-            <p><strong>شماره نظام پزشکی:</strong> {{ doctor.license_number }} </p>
-        </div>
-
-        <div class="about-doctor">
-            <h2>درباره دکتر</h2>
-            <p>{{ doctor.biography }}</p>
-        </div>
-
-        <div class="doctor-status">
-            {% if doctor.is_active %}
-                <p class="status active">دکتر درحال حاضر پذیرش دارد</p>
-            {% else %}
-                <p class="status not-active">دکتر درحال حاضر پذیرش ندارد</p>
-            {% endif %}
-        </div>
-
-       <div class="visit-cost">
-            <p><strong>هزینه ویزیت:</strong> {{ doctor.visit_cost }} ریال</p>
-            <a href="{% url 'availabletime_doctor' doctor.id %}">نوبت دهی</a>
-        </div>
-
-<div class="comments-section">
-    {% with comments.count as total_comments %}
-        <h2>
-          تعداد نظرات: {{ total_comments }}
-        </h2>
-    {% endwith %}
-
-    {% for comment in comments %}
-        <div class="comment">
-            <p class="comment-author">نظر توسط {{ comment.patient.user.username }}</p>
-            <div class="comment-text">{{ comment.description|linebreaks }}</div>
-            <p class="comment-date">در تاریخ {{ comment.created_at }}</p>
-        </div>
-    {% empty %}
-        <p>هیچ نظری ثبت نشده است</p>
-    {% endfor %}
-    <hr>
-    <div class="new-comment-form">
-        {% if user.is_authenticated %}
-            <h2>نظر خود را بنویسید</h2>
-            <form method="post">
-                {% csrf_token %}
-                {{ comment_form.as_p }}
-                <input type="submit" value="ثبت نظر" class="submit-btn">
-            </form>
-        {% else %}
-            <p>برای ثبت نظر باید وارد حساب کاربری خود شوید.</p>
-        {% endif %}
-    </div>
-</div>
-</body>
-
-<script>
-    // انتخاب پیام‌ها
-    document.querySelectorAll('.messages li').forEach(function(message) {
-        // ناپدید شدن پیام‌ها بعد از ۳ ثانیه
-        setTimeout(function() {
-            message.style.opacity = '0'; // به تدریج ناپدید می‌شود
-            setTimeout(function() {
-                message.remove(); // حذف کامل پیام از DOM
-            }, 1000); // زمان تطابق با transition
-        }, 3000); // 3 ثانیه برای نمایش
-    });
-</script>
-{% endblock %}
