+++ management/templates/patient/add_patient.html
