+++ management/templates/patient/patient_reserved_times.html
@@ -49,7 +49,7 @@
                 setTimeout(function() {
                     message.remove(); // حذف کامل پیام از DOM
                 }, 1000); // زمان تطابق با transition
-            }, 3000); // 3 ثانیه برای نمایش
+            }, 5000); // 3 ثانیه برای نمایش
         });
     </script>
 {% endblock %}
