--- management/templates/reset_password.html
@@ -1,43 +1,0 @@
-
-
-
-{% extends 'base.html' %}
-{% load static %}
-{% block content %}
-    {% block extra_css %}
-    <link rel="stylesheet" href="{% static 'css/create_doctor.css' %}">
-    {% endblock %}
-    <div class="container">
-        <h1 class="page-title">تغییر پسورد</h1>
-
-        <div class="message-container">
-            {% if messages %}
-            <ul class="messages">
-                {% for message in messages %}
-                    <li class="{{ message.tags }}" id="message-{{ forloop.counter }}">{{ message }}</li>
-                {% endfor %}
-            </ul>
-            {% endif %}
-        </div>
-        
-        <div class="form-container">
-            <form method="post">
-                {% csrf_token %}
-                {{ form.as_p }}
-                <button type="submit" class="submit-btn">تغییر رمز عبور</button>
-            </form>
-        </div>
-    </div>
-
-    <script>
-        // ناپدید شدن پیام‌ها بعد از ۳ ثانیه
-        document.querySelectorAll('.messages li').forEach(function(message) {
-            setTimeout(function() {
-                message.style.opacity = '0'; // به تدریج ناپدید می‌شود
-                setTimeout(function() {
-                    message.remove(); // حذف کامل پیام از DOM
-                }, 1000); // زمان تطابق با transition
-            }, 3000); // 3 ثانیه برای نمایش
-        });
-    </script>
-{% endblock %}
