+++ management/urls.py
@@ -1,12 +1,15 @@
 from django.urls import path
-from .views import *
+from management.views.av_time_view import *
+from management.views.doctor_view import *
+from management.views.patient_view import *
+from management.views.login_view import *
+
 
 urlpatterns = [
     path('create_patient/', create_patient, name='create_patient'),
     path('createdoctor/', create_doctor, name='createdoctor'),
     path('viewdoctor/', view_doctor, name='viewdoctor'),
     path('createspecialize/', create_specialize, name='createspecialize'),
-    path('create_rating/', create_rating, name='create_rating'),
     path('delete_doctor/<int:id>/', delete_doctor, name='delete_doctor'),
     path('detail_doctor/<int:id>/', detail_doctor, name='detail_doctor'),
     path('edit_doctor/<int:id>/', edit_doctor, name='edit_doctor'),
@@ -24,4 +27,7 @@ urlpatterns = [
     path('forgot_password/', forgot_password_view, name="forgot_password"),
     path('reset_password/', reset_password_view, name="reset_password"),
     path('patient_profile', patient_profile, name='patient_profile'),
+    path('delete-comment/<int:doctor_id>/<int:comment_id>/', delete_comment, name='delete_comment'),
+    # path('doctor/add_rating/<int:doctor_id>/', add_rating, name='add_rating'),
+
 ]
