+++ static/css/availabletime_doctor.css
@@ -79,32 +79,32 @@ h1 {
 
 
 
-    /* استایل پیام‌ها */
-    .message-container {
-        margin-bottom: 20px;
-    }
+/* استایل پیام‌ها */
+.message-container {
+    margin-bottom: 20px;
+}
 
-    .messages {
-        list-style: none;
-        padding: 0;
-    }
+.messages {
+    list-style: none;
+    padding: 0;
+}
 
-    .messages li {
-        padding: 15px;
-        margin-bottom: 15px; /* فاصله بیشتر بین پیام‌ها */
-        border-radius: 5px;
-        font-size: 16px; /* افزایش اندازه فونت برای خوانایی بهتر */
-        transition: opacity 1s ease;
-        display: flex;
-        align-items: center; /* تراز عمودی */
-    }
+.messages li {
+    padding: 15px;
+    margin-bottom: 15px; /* فاصله بیشتر بین پیام‌ها */
+    border-radius: 5px;
+    font-size: 16px; /* افزایش اندازه فونت برای خوانایی بهتر */
+    transition: opacity 1s ease;
+    display: flex;
+    align-items: center; /* تراز عمودی */
+}
 
-    .messages li.success {
-        background-color: #2ecc71;
-        color: white;
-    }
+.messages li.success {
+    background-color: #2ecc71;
+    color: white;
+}
 
-    .messages li.error {
-        background-color: #e74c3c;
-        color: white;
-    }
+.messages li.error {
+    background-color: #e74c3c;
+    color: white;
+}
